feat: move download logic to download module
