+++ src/main.rs
@@ -357,45 +357,6 @@ async fn main() -> anyhow::Result<()> {
         }
         Commands::Download { output, torrent } => {
             let t = Torrent::read(torrent).await?;
-            let info_hash = t.info_hash();
-            let req = tracker::http::Request::new(&info_hash, t.length());
-            let url = req.url(&t.announce);
-
-            let res = reqwest::get(url).await?;
-            let res_bytes = res.bytes().await?;
-            let tracker_res: tracker::http::Response = serde_bencode::from_bytes(&res_bytes)?;
-
-            let shared_state = Arc::new(SharedState::new(t.info.pieces.0.len()));
-
-            let futures = tracker_res
-                .peers
-                .0
-                .into_iter()
-                .map(|addr| {
-                    let info_hash = info_hash.clone();
-                    async move { Peer::new(addr, &info_hash).await }
-                })
-                .collect::<Vec<_>>();
-
-            let results = join_all(futures).await;
-            for result in results {
-                match result {
-                    Ok(mut connection) => {
-                        let piece =
-                            download_worker(&t, shared_state.clone(), &mut connection).await;
-                    }
-                    Err(e) => eprintln!("Error: {}", e),
-                };
-            }
-
-            let all_pieces = shared_state.all_pieces();
-            let mut output_file = tokio::fs::File::options()
-                .write(true)
-                .create(true)
-                .open(&output)
-                .await
-                .unwrap();
-            output_file.write_all(&all_pieces).await?;
 
             println!("Downloaded test.torrent to {}.", output.display());
         }
