+++ src/tracker/http.rs
@@ -77,6 +77,9 @@ impl Response {
             peers: Peers(Vec::new()),
         }
     }
+
+    async fn query() -> anyhow::Result<Self> {
+    }
 }
 
 #[derive(Debug, Clone)]
